add Client.fetch and simplify things
