+++ api/client.go
@@ -209,3 +209,26 @@ func (c *Client) doRequest(request *http.Request) ([]byte, error) {
 
 	return body, nil
 }
+
+func (c *Client) fetch(path string) (bool, []byte, error) {
+	req, err := http.NewRequest("GET", c.APIEndpoint+path, nil)
+	if err != nil {
+		return false, nil, err
+	}
+
+	response, err := c.doRequestRaw(req)
+	if err != nil {
+		return false, nil, err
+	}
+
+	body, err := ioutil.ReadAll(response.Body)
+	if err != nil {
+		return false, nil, err
+	}
+
+	if response.StatusCode != 200 {
+		return false, body, nil
+	}
+
+	return true, body, nil
+}
